+++ frontend/frontendserver.py
@@ -1,0 +1,8 @@
+import http.server
+import socketserver
+
+PORT = 8000
+handler = http.server.SimpleHTTPRequestHandler
+
+with socketserver.TCPServer(("", PORT), handler) as httpd:
+    httpd.serve_forever()
