+++ frontend/index.html
@@ -9,10 +9,11 @@
     <div id="workspace-wrapper">
         <canvas id="workspace"></canvas>
     </div>
-    <img crossOrigin="Anonymous" class="imag-res" id="img-logo" src="images/logo.svg">
-    <script src="js/minMatrix.js"></script>
-    <script src="js/shaders.js"></script>
-    <script src="js/webgl.js"></script>
+    <img crossOrigin="anonymous" id="img-logo" src="images/logo.png">
+    <script src="js/webgl/matrix.js"></script>
+    <script src="js/webgl/model.js"></script>
+    <script src="js/webgl/shaders.js"></script>
+    <script src="js/webgl/webgl.js"></script>
     <script src="js/main.js"></script>
 </body>
